Changed datatype from int to double, removoed obosolete test case, removed duplicate exception handler, added sql data file for tests that was accidentially deleted

+++ server/src/main/java/com/tablefootbal/server/dto/SensorDto.java
@@ -6,5 +6,5 @@ import lombok.Getter;
 public class SensorDto
 {
 	String id;
-	int[] readings;
+	double[] readings;
 }
